Move fused RTL scheduling into celox-slt
