+++ crates/celox-slt/Cargo.toml
@@ -8,12 +8,14 @@ description           = "Source-independent symbolic logic tree core for Celox"
 edition.workspace     = true
 
 [dependencies]
+celox-analysis = { path = "../celox-analysis" }
 celox-design = { path = "../celox-design" }
 celox-sir    = { path = "../celox-sir" }
 fxhash       = { workspace = true }
 num-bigint   = { workspace = true }
 num-traits   = { workspace = true }
 serde        = { workspace = true }
+thiserror     = { workspace = true }
 
 [dev-dependencies]
 serde_json = { workspace = true }
